test: split DBF memo integrity regressions
